Move more logic to DrinkDetailsCard
